Move CLI commands to namespace
