+++ reward_index/src/cli/mod.rs
@@ -1,0 +1,2 @@
+pub mod escrow;
+pub mod server;
